Fix segfault and other problems when broken frames are encountered.
The current buffer sizes do not allow 5.1 channel audio frames to
be completely accommodated.  By their nature, the Softmixer and
Equalizer operate on whole frames and will fail if they receive a
partial frame and the end of the buffer.

This is a workaround rather than a complete fix because we should
ensure that the decoders only ever return complete frames, but this
is difficult to do when the frame size may change mid-stream.

git-svn-id: svn://daper.net/moc/trunk@2582 910807d9-36e0-0310-a014-e9ea483e2ba4