CUE parsing: First pass with libcue
CUE files do get processed, but libcue has problems with some of them.
Do not expect CUE functionality to work well at this point. There are
concerns with how the title tags are created with CUE files, the logic
seems to go against the design of MOC. To be continued.