C99: Silence data to function pointer conversion warnings.
It turns out that both of the previous approaches to this problem are
correct, but in different contexts.  r2797 is correct when compiling
with ANSI C99 while r2825 is correct when compiling with GNU's C99.

However, they're only correct in terms of compiler warnings; in fact,
neither are correct because there is *no* correct way of doing this.
The origin of the problem lies in a conflict between the C99 and
POSIX standards.

C99 says (correctly) that assigning a data pointer to a function
pointer is undefined.  For some machine architectures, the two
pointer types may represented differently.  The POSIX standard says
that dlsym(3) returns the address of a symbol as a data pointer.
If that symbol is a function then there is no legal way of assigning
it to a function pointer.

On most modern machine architectures, and certainly most on which
MOC will ever run, the two pointer types are identical and can be
used interchangably.  Therefore, the approach taken in this patch
assumes that they are and avoids compiler warnings via a union.
We check that function and data pointers are the same length and
warn the user that memory accessing or illegal instruction errors may
result if they're not or if we can't determine that they are.  Even if
they are the same length they may still be represented differently,
but there seems no way of checking for this that does not involve
doing a trial conversion.

In summary, this approach works as well as any other and will avoid
warnings about which nothing can be done.  In most cases it will work;
in a few it will fail, and there's no general solution.

This will be revisited when the plug-in architecture is overhauled.

Amends: r2797 and r2825

git-svn-id: svn://daper.net/moc/trunk@2831 910807d9-36e0-0310-a014-e9ea483e2ba4