C99: Work around Berkeley DB's header generation assumptions.
The configuration of Berkeley DB itself conditionally generates
typedefs for the u_* types within the db.h header.  That doesn't
work if they were not included in db.h but the client application
is configured in such a way which causes them not to be exposed by
sys/types.h at build time.

git-svn-id: svn://daper.net/moc/trunk@2838 910807d9-36e0-0310-a014-e9ea483e2ba4