Fix: Use fcntl(2) rather than lockf(3) for playlist file locking.
On Linux (at least) lockf(3) uses fcntl(3) for locking anyway, but
always uses write lock even if the file is opened read-only.  This
results in the file not being locked when reading a playlist and a
"Can't lock the playlist file: Bad file descriptor" error being
issued.

(Earlier lockf(3) manpages did not mention that the file must be
writable.)

Amends: r2834

git-svn-id: svn://daper.net/moc/trunk@2943 910807d9-36e0-0310-a014-e9ea483e2ba4